requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==1.0.0
Django==4.1
django-filter==22.1
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
mysqlclient==2.1.1
psycopg2==2.9.3
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
django-cors-headers==3.13.0